+++ src/librt/primitives/nmg/nmg_ck.c
@@ -683,7 +683,7 @@ nmg_ck_e(const struct edgeuse *eu, const struct edge *e, const char *str)
 {
     char *errstr;
     struct edgeuse *eparent;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_e error str");
     snprintf(errstr, len, "%sedge %8lx\n", str, (unsigned long)e);
@@ -717,7 +717,7 @@ void
 nmg_ck_vu(const unsigned long *parent, const struct vertexuse *vu, const char *str)
 {
     char *errstr;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_vu error str");
     snprintf(errstr, len, "%svertexuse %8lx\n", str, (unsigned long)vu);
@@ -739,7 +739,7 @@ nmg_ck_eu(const unsigned long *parent, const struct edgeuse *eu, const char *str
 {
     char *errstr;
     struct edgeuse *eur, *eu_next, *eu_last;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_eu error str");
     snprintf(errstr, len, "%sedgeuse %8lx\n", str, (unsigned long)eu);
@@ -819,7 +819,7 @@ void
 nmg_ck_lg(const struct loop *l, const struct loop_g *lg, const char *str)
 {
     char *errstr;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_lg error str");
     snprintf(errstr, len, "%sloop_g %8lx\n", str, (unsigned long)lg);
@@ -838,7 +838,7 @@ void
 nmg_ck_l(const struct loopuse *lu, const struct loop *l, const char *str)
 {
     char *errstr;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_l error str");
     snprintf(errstr, len, "%sloop %8lx\n", str, (unsigned long)l);
@@ -869,7 +869,7 @@ nmg_ck_lu(const unsigned long *parent, const struct loopuse *lu, const char *str
     int l;
     int edgeuse_num=0;
     unsigned long magic1;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_lu error str");
     snprintf(errstr, len, "%sloopuse %8lx\n", str, (unsigned long)lu);
@@ -913,7 +913,7 @@ nmg_ck_lu(const unsigned long *parent, const struct loopuse *lu, const char *str
 	NMG_CK_VERTEXUSE(vu);
 	nmg_ck_vu(&lu->l.magic, vu, errstr);
     } else if (magic1 == NMG_EDGEUSE_MAGIC) {
-	l = strlen(errstr);
+	l = (int)strlen(errstr);
 	for (BU_LIST_FOR(eu, edgeuse, &lu->down_hd)) {
 	    NMG_CK_EDGEUSE(eu);
 	    snprintf(&errstr[l], len-l, "%sedgeuse #%d (%8lx)\n",
@@ -935,7 +935,7 @@ void
 nmg_ck_fg(const struct face *f, const struct face_g_plane *fg, const char *str)
 {
     char *errstr;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_fg error str");
     snprintf(errstr, len, "%sFace_g %8lx\n", str, (unsigned long)f);
@@ -959,7 +959,7 @@ void
 nmg_ck_f(const struct faceuse *fu, const struct face *f, const char *str)
 {
     char *errstr;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     errstr = bu_calloc(len, 1, "nmg_ck_f error str");
     snprintf(errstr, len, "%sFace %8lx\n", str, (unsigned long)f);
@@ -988,7 +988,7 @@ nmg_ck_fu(const struct shell *s, const struct faceuse *fu, const char *str)
     int l;
     int loop_number = 0;
     struct loopuse *lu;
-    int len = strlen(str)+128;
+    int len = (int)strlen(str)+128;
 
     NMG_CK_FACEUSE(fu);
     NMG_CK_SHELL(s);
@@ -1024,7 +1024,7 @@ nmg_ck_fu(const struct shell *s, const struct faceuse *fu, const char *str)
 
     nmg_ck_f(fu, fu->f_p, errstr);
 
-    l = strlen(errstr);
+    l = (int)strlen(errstr);
     for (BU_LIST_FOR(lu, loopuse, &fu->lu_hd)) {
 	NMG_CK_LOOPUSE(lu);
 	snprintf(&errstr[l], len-l, "%sloopuse #%d (%8lx)\n",
